pyproject.toml
[tool.poetry]
name = "course-work-7"
version = "0.1.0"
description = ""
authors = ["DaniilKokorev <kokorevdaniil@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
ipython = "^8.26.0"
getenv = "^0.2.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
django-celery-beat = "^2.6.0"
celery = "^5.4.0"
redis = "^5.0.8"
gevent = "^24.2.1"
requests = "^2.32.3"
coverage = "^7.6.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
